requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==6.7
dnspython==2.0.0
email-validator==1.1.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.10
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
newsapi-python==0.2.6
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
requests==2.24.0
six==1.15.0
SQLAlchemy==1.2.8
urllib3==1.25.10
Werkzeug==0.14.1
WTForms==2.2.1
